+++ reporting/rep201.php
@@ -164,7 +164,7 @@ function print_supplier_balances(): void
 		if ((bool)$show_balance)
 		{
 			$init[3] = $init[0] - $init[1];
-			$accumulate += (float)$init[3];
+			$accumulate += $init[3];
 		}	
 		else	
 			$init[3] = round2(($bal != false ? $bal['OutStanding'] : 0)*$rate, $dec);
@@ -183,8 +183,8 @@ function print_supplier_balances(): void
 		$total = array(0,0,0,0);
 		for ($i = 0; $i < 4; $i++)
 		{
-			$total[$i] += (float)$init[$i];
-			$grandtotal[$i] += (float)$init[$i];
+			$total[$i] += $init[$i];
+			$grandtotal[$i] += $init[$i];
 		}
 		$rep->NewLine(1, 2);
 		$rep->Line($rep->row + 4.0);
@@ -206,15 +206,15 @@ function print_supplier_balances(): void
 			{
 				$item[0] = round2(abs($trans['TotalAmount']) * $rate, $dec);
 				$rep->AmountCol(4, 5, $item[0], $dec);
-				$accumulate += (float)$item[0];
+				$accumulate += $item[0];
 				$item[2] = round2((float)$trans['Allocated'] * (float)$rate, $dec);
 			}
 			else
 			{
 				$item[1] = round2(abs($trans['TotalAmount']) * $rate, $dec);
 				$rep->AmountCol(5, 6, $item[1], $dec);
-				$accumulate -= (float)$item[1];
-				$item[2] = (float)round2((float)$trans['Allocated'] * (float)$rate, $dec) * (float)(-1);
+				$accumulate -= $item[1];
+				$item[2] = round2((float)$trans['Allocated'] * (float)$rate, $dec) * (float)(-1);
 			}
 			$rep->AmountCol(6, 7, $item[2], $dec);
 			if ($trans['TotalAmount'] > 0.0)
