+++ reporting/rep206.php
@@ -160,15 +160,15 @@ function print_supplier_balances(): void
         $init[2] = round2(($bal != false ? $bal['Allocated'] : 0)*$rate, $dec);
 
         $init[3] = $init[1] - $init[0];
-        $accumulate += (float)$init[3];
+        $accumulate += $init[3];
 
         $res = getTransactions($myrow['supplier_id'], $from, $to);
 
         $total = array(0,0,0,0);
         for ($i = 0; $i < 4; $i++)
         {
-            $total[$i] += (float)$init[$i];
-            $grandtotal[$i] += (float)$init[$i];
+            $total[$i] += $init[$i];
+            $grandtotal[$i] += $init[$i];
         }
 
         if (db_num_rows($res) == 0 && !(bool)$no_zeros) 
@@ -189,23 +189,23 @@ function print_supplier_balances(): void
             if ($trans['TotalAmount'] > 0.0)
             {
                 $item[0] = round2(abs($trans['TotalAmount']) * $rate, $dec);
-                $curr_cr += (float)$item[0];
-                $tot_cur_cr += (float)$item[0];
+                $curr_cr += $item[0];
+                $tot_cur_cr += $item[0];
 
                 $accumulate -= $item[0];
             }
             else
             {
                 $item[1] = round2(abs($trans['TotalAmount']) * $rate, $dec);
-                $curr_db += (float)$item[1];
-                $tot_cur_db += (float)$item[1];
+                $curr_db += $item[1];
+                $tot_cur_db += $item[1];
                 $accumulate += $item[1];
             }
             $item[2] = round2((float)$trans['Allocated'] * (float)$rate, $dec);
             if ($trans['TotalAmount'] > 0.0)
                 $item[3] = $item[2] - $item[0];
             else
-                $item[3] = ((float)($item[2] - $item[1])) * (float)(-1);
+                $item[3] = (($item[2] - $item[1])) * (float)(-1);
 
             for ($i = 0; $i < 4; $i++)
             {
@@ -216,7 +216,7 @@ function print_supplier_balances(): void
         }
 		if ((bool)$no_zeros && $total[3] == 0.0 && $curr_db == 0.0 && $curr_cr == 0.0) continue;
         $rep->TextCol(0, 2, (string)$myrow['name'].($myrow['inactive']==1 ? " ("._("Inactive").")" : ""));
-        $rep->AmountCol(3, 4, (float)$total[3] + (float)$curr_cr - $curr_db, $dec);
+        $rep->AmountCol(3, 4, (float)$total[3] + $curr_cr - $curr_db, $dec);
         $rep->AmountCol(4, 5, $curr_db, $dec);
         $rep->AmountCol(5, 6, $curr_cr, $dec);
         $rep->AmountCol(7, 8, $total[3], $dec);
